Discard router broadcast packets which are received on the wrong multicast address
This allows promiscuous mode to be turned on (e.g. tcpdump) with changing the
operation of the DECnet stack. Previously endnode hello messages would be
received and processed populating the neighbour cache and modifying the
operation of the first few messages of connection establishment. Also added
the gateway address to /proc/net/decnet_cache to make debuggibg easier.