Fix net_device reference count when inbound packet constructs a new route.
There are 2 typical cases where a route is constructed during inbound packet
processing (in dn_route_input_slow()):

1. A new connection is being requested from a node we've never seen before.

2. A connect confirm is received from a different node than the one we sent
   the original connect initiate to. This can occur when the destination
   node is on the ethernet but we did not know about it and sent the
   connect initiate to the designated router and the connect confirm came
   back directly from the destination with the "intra-ethernet" bit set.

There are a number of other cases, but these are the most likely. The code
built a new route entry using the input device but did not update the
reference count on the device.