Use standard sock_queue_rcv_skb() rather than custom dn_queue_skb().
dn_queue_skb() was a custom routine derived from sock_queue_rcv_skb() without
a call to bh_lock_sock(). It is unclear when this derivation occurred and,
since then, none of the changes to sock_queue_rcv_skb() have been merged in.
Replace 2 calls to the custom routine with calls to the standard routine
with the appropriate rcu locking code.