Add/check timestamp to route table entries.
If a node is started with no desinated router on the LAN and an attempt is made
to connect to a remote node which would normally go through the router, the
connection will fail and the invalid route entry will remain active for two
minutes. If the designated router comes up within those 2 minutes and an
attempt is made to connect to the remote node it will fail and the invalid
route entry will stay around for an additional 2 minutes. This problem may
also occur if the designated router changes or the current router times out.
Add a timestamp to each route entry and a global timestamp for when the
designated router changes. When looking up a route, ignore entries which were
created before the last designated router state change.